Fix: 투표 - 매매 로직 연동 개선

+++ vote-service/src/main/java/com/example/vote_service/dto/payload/TradeExecutedPayloadDTO.java
@@ -6,7 +6,7 @@ package com.example.vote_service.dto.payload;
 public record TradeExecutedPayloadDTO(
         String side,              // "BUY" | "SELL"
         String stockName,
-        Integer shares,
+        Float shares,             // 소수점 거래 지원을 위해 Float으로 변경
         Integer unitPrice,
         Integer accountBalance    // 체결 후 모임계좌잔액
 ) {
