+++ vote-service/src/main/java/com/example/vote_service/dto/payload/TradePayload.java
@@ -18,7 +18,7 @@ public record TradePayload(
         @NotBlank String reason,                    // 제안 이유
         @NotNull @JsonProperty("stockId") UUID stockId,        // 주식 ID (JSON: stockId, DB: stock_id)
         @NotBlank @JsonProperty("stockName") String stockName, // 주식명 (JSON: stockName, DB: stock_name)
-        @NotNull @Positive Integer price,           // 주가
-        @NotNull @Positive Integer quantity         // 수량
+        @NotNull @Positive Integer price,           // 주가 (Integer 유지 - 원 단위)
+        @NotNull @Positive Float quantity           // 수량 (Float으로 변경 - 소수점 거래 지원)
 ) {
 }
